feat: implement logout functionality in ProfilePage and remove from Header
